// frontend/src/app/models/DataDto.ts
export default class DataDto {
    Embarazos: number;
    Glucosa: number;
    PreTrueonSanguinea: number;
    PliegueCutaneo: number;
    Insulina: number;
    IndiceMasaCorporal: number;
    PedigriDiabetes: number;
    Edad: number;
    MedicacionPrevia: string = "nan";
    Observaciones: string = "nan";
    FechaDiagostico: string = "2016-03-04 00:00:00,000000";
}